Fix Cursor Rendering and Improper Gap Movement
Fix #3

The visual bug of cursor begin rendered one square off after the gap
was caused as a result of gap_start begin included in the column
number calculation of the cursor.

The bug was also a result of improper gap movement and relocation of
chars after the gap, which prior to this fix was also moving the char
at the cursor which inserted the char one off to the right.